pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mindrouter2"
version = "0.19.1"
description = "Production-ready LLM inference load balancer for Ollama and vLLM backends"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "MindRouter Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["llm", "load-balancer", "ollama", "vllm", "inference", "ai"]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.0",
    "pymysql>=1.1.0",
    "aiomysql>=0.2.0",
    "cryptography>=42.0.0",
    "redis>=5.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[argon2]>=1.7.4",
    "argon2-cffi>=23.1.0",
    "jinja2>=3.1.0",
    "tiktoken>=0.5.0",
    "jsonschema>=4.20.0",
    "prometheus-client>=0.19.0",
    "structlog>=24.1.0",
    "aiofiles>=23.2.0",
    "itsdangerous>=2.1.0",
    "starlette-session>=0.4.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "pdfplumber>=0.10.0",
    "beautifulsoup4>=4.12.0",
    "authlib>=1.3.0",
    "markdown>=3.5.0",
    "pygments>=2.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
    "black>=24.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "types-redis>=4.6.0",
    "sqlalchemy-stubs>=0.4",
    "faker>=22.0.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "faker>=22.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/mindrouter2"
Documentation = "https://github.com/yourusername/mindrouter2#readme"
Repository = "https://github.com/yourusername/mindrouter2"
Issues = "https://github.com/yourusername/mindrouter2/issues"

[project.scripts]
mindrouter = "backend.app.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

[tool.pytest.ini_options]
testpaths = ["backend/app/tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]
addopts = "-v --tb=short"
markers = [
    "integration: tests that hit real backend services (Ollama, vLLM)",
]

[tool.coverage.run]
source = ["backend/app"]
omit = [
    "backend/app/tests/*",
    "backend/app/db/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["backend"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["sqlalchemy.ext.mypy.plugin"]
